Move test `common` module to `mullvad-tests` lib.
