+++ mullvad-tests/tests/startup.rs
@@ -1,15 +1,11 @@
-#[macro_use]
-extern crate duct;
-extern crate mullvad_ipc_client;
 extern crate mullvad_paths;
-
-mod common;
+extern crate mullvad_tests;
 
 use std::fs::{self, Metadata};
 use std::io;
 use std::time::Duration;
 
-use common::DaemonRunner;
+use mullvad_tests::DaemonRunner;
 
 use platform_specific::*;
